feat: creation of account collector and move of related functions
